Update and rename NumberToPair.cs to NumberToColor.cs
